MySqlDocker/requirement.txt
streamlit==1.21.0
recordlinkage==0.15
pandas==1.5.3
streamlit-aggrid==0.2.2.post2
preprocessing==0.1.13
grouping==0.1.8
Pillow==9.5.0
numpy==1.24.3
cryptography==40.0.2
geopy==2.3.0
pathlib==1.0.1
uuid==1.30
vl-convert-python==0.10.2
splink==3.8.1
pymysql==1.0.3
PyYAML==6
openpyxl==3.1.2
sqlalchemy==2.0.15
